requirements.txt
pixivpy3
python-telegram-bot
aiohttp